add sc secret

+++ src/main/java/com/jq/findapp/api/SupportCenterApi.java
@@ -73,15 +73,23 @@ public class SupportCenterApi {
 	@Value("${app.scheduler.secret}")
 	private String schedulerSecret;
 
+	@Value("${app.supportCenter.secret}")
+	private String supportCenterSecret;
+
 	@DeleteMapping("user/{id}")
 	public void userDelete(@PathVariable final BigInteger id, @RequestHeader String password,
-			@RequestHeader String salt) throws Exception {
+			@RequestHeader String salt, @RequestHeader String secret) throws Exception {
+		if (!supportCenterSecret.equals(secret))
+			return;
 		authenticationService.verify(adminId, password, salt);
 		authenticationService.deleteAccount(repository.one(Contact.class, id));
 	}
 
 	@GetMapping("user")
-	public List<Object[]> user(@RequestHeader String password, @RequestHeader String salt) {
+	public List<Object[]> user(@RequestHeader String password, @RequestHeader String salt,
+			@RequestHeader String secret) {
+		if (!supportCenterSecret.equals(secret))
+			return null;
 		final QueryParams params = new QueryParams(Query.contact_listSupportCenter);
 		params.setUser(authenticationService.verify(adminId, password, salt));
 		params.setLimit(Integer.MAX_VALUE);
@@ -89,7 +97,8 @@ public class SupportCenterApi {
 	}
 
 	@GetMapping("ticket")
-	public List<Object[]> ticket(String search, @RequestHeader String password, @RequestHeader String salt) {
+	public List<Object[]> ticket(String search, @RequestHeader String password, @RequestHeader String salt,
+			@RequestHeader String secret) {
 		final QueryParams params = new QueryParams(Query.misc_listTicket);
 		params.setUser(authenticationService.verify(adminId, password, salt));
 		params.setSearch(search);
@@ -99,12 +108,13 @@ public class SupportCenterApi {
 
 	@DeleteMapping("ticket/{id}")
 	public void ticketDelete(@PathVariable final BigInteger id, @RequestHeader String password,
-			@RequestHeader String salt) throws Exception {
+			@RequestHeader String salt, @RequestHeader String secret) throws Exception {
 		repository.delete(repository.one(Ticket.class, id));
 	}
 
 	@GetMapping("log")
-	public List<Object[]> log(String search, @RequestHeader String password, @RequestHeader String salt) {
+	public List<Object[]> log(String search, @RequestHeader String password, @RequestHeader String salt,
+			@RequestHeader String secret) {
 		final QueryParams params = new QueryParams(Query.misc_listLog);
 		params.setUser(authenticationService.verify(adminId, password, salt));
 		params.setSearch(search);
@@ -115,14 +125,14 @@ public class SupportCenterApi {
 	@PostMapping("email")
 	@Produces(MediaType.TEXT_PLAIN)
 	public void email(final BigInteger id, final String text, final String action,
-			@RequestHeader String password, @RequestHeader String salt) throws Exception {
+			@RequestHeader String password, @RequestHeader String salt, @RequestHeader String secret) throws Exception {
 		final Contact contact = authenticationService.verify(adminId, password, salt);
 		notificationService.sendNotificationEmail(contact, repository.one(Contact.class, id), text, action);
 	}
 
 	@PostMapping("chat")
 	public void chat(@RequestBody Notification data, @RequestHeader String password,
-			@RequestHeader String salt) throws Exception {
+			@RequestHeader String salt, @RequestHeader String secret) throws Exception {
 		authenticationService.verify(adminId, password, salt);
 		for (BigInteger id : data.getIds()) {
 			final Chat chat = new Chat();
@@ -135,7 +145,8 @@ public class SupportCenterApi {
 	}
 
 	@PutMapping("resend/{id}")
-	public void resend(@PathVariable final BigInteger id, @RequestHeader String password, @RequestHeader String salt)
+	public void resend(@PathVariable final BigInteger id, @RequestHeader String password, @RequestHeader String salt,
+			@RequestHeader String secret)
 			throws Exception {
 		authenticationService.verify(adminId, password, salt);
 		authenticationService.recoverSendEmailReminder(repository.one(Contact.class, id));
@@ -143,7 +154,7 @@ public class SupportCenterApi {
 
 	@PutMapping("log/search")
 	public void logSearch(@RequestBody List<Map<String, Object>> searches, @RequestHeader String password,
-			@RequestHeader String salt) throws Exception {
+			@RequestHeader String salt, @RequestHeader String secret) throws Exception {
 		authenticationService.verify(adminId, password, salt);
 		final QueryParams param = new QueryParams(Query.misc_setting);
 		param.setSearch("setting.label='sc.search'");
@@ -154,7 +165,8 @@ public class SupportCenterApi {
 	}
 
 	@GetMapping("log/search")
-	public String logSearch(@RequestHeader String password, @RequestHeader String salt) throws Exception {
+	public String logSearch(@RequestHeader String password, @RequestHeader String salt, @RequestHeader String secret)
+			throws Exception {
 		authenticationService.verify(adminId, password, salt);
 		final QueryParams param = new QueryParams(Query.misc_setting);
 		param.setSearch("setting.label='sc.search'");
